feat: intagrate rabbitmq
enable roles and their check

+++ app/roles/owner/admins_management/admins_management.py
@@ -10,18 +10,20 @@ from aiogram.types import (
 
 from app.config import labels
 from app.config.config import OWNERS
-from app.database.admin_db_operations import (
-    add_admin_to_db,
-    delete_admin_from_db,
-    get_admin_by_id,
+from app.database.user_db_operations import (
+    add_or_update_user_to_admin,
+    demote_admin_to_user,
+    get_user_by_id,
+    get_user_by_telegram_tag,
 )
 from app.keyboards import (
     inline_admin_list,
     inline_single_cancel_button,
     main_actions_keyboard,
 )
+from app.roles.owner.owner import owner  # Импортируем роутер owner
 from app.roles.user.callbacks_enum import Callbacks
-from app.roles.user.user_cmds import logger, user
+from app.utils.logger import logger
 
 
 class AdminManagementStates(StatesGroup):
@@ -29,26 +31,22 @@ class AdminManagementStates(StatesGroup):
 
 
 def is_valid_telegram_username(username: str) -> bool:
-    """
-    Validates Telegram-username.
-    Username must:
-    - Start with @
-    - Have 5 or more symbols
-    - No digit or _ after @
-    - Have only letters (a-z, A-Z), digits (0-9) and _
-    """
     return (
-        username.startswith("@")
-        and len(username) >= 5
-        and not username[1].isdigit()
-        and username[1] != "_"
-        and all(c.isalnum() or c == "_" for c in username[1:])
+            username.startswith("@")
+            and len(username) >= 5
+            and not username[1].isdigit()
+            and username[1] != "_"
+            and all(c.isalnum() or c == "_" for c in username[1:])
     )
 
 
-@user.message(F.text == labels.MANAGE_ADMINS)
+@owner.message(F.text == labels.MANAGE_ADMINS)
 async def manage_admins(message: Message):
     logger.info("manage_admins_call")
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     await message.answer(
         text="Выберите админа или создайте нового:",
         reply_markup=await inline_admin_list(
@@ -65,19 +63,19 @@ async def on_add_admin_clicked(callback: CallbackQuery, state: FSMContext):
     await callback.answer()
 
 
-@user.callback_query(F.data == Callbacks.add_admin_callback)
+@owner.callback_query(F.data == Callbacks.add_admin_callback)
 async def handle_add_admin_callback(callback: CallbackQuery, state: FSMContext):
     await on_add_admin_clicked(callback, state)
 
 
-@user.message(AdminManagementStates.waiting_for_admin_username)
+@owner.message(AdminManagementStates.waiting_for_admin_username)
 async def process_admin_username(message: Message, state: FSMContext):
     username = message.text.strip()
     if not is_valid_telegram_username(username):
         await message.answer(
-            text="Некорректный формат! Используйте '@username'\
-(минимум 5 символов, только буквы, цифры и _, "
-            "не начинается с цифры или _).\nВведите снова:",
+            text="Некорректный формат! Используйте '@username'\n"
+                 "(минимум 5 символов, только буквы, цифры и _, "
+                 "не начинается с цифры или _).\nВведите снова:",
             reply_markup=await inline_single_cancel_button(
                 Callbacks.cancel_primary_action_callback
             ),
@@ -85,13 +83,17 @@ async def process_admin_username(message: Message, state: FSMContext):
         return
     if username in OWNERS:
         await message.answer(
-            text=f"Админ '{username}' не может быть добавлен, так как он уже является владельцем!",
+            text=f"Пользователь '{username}' является владельцем и не может быть изменён!",
             reply_markup=await inline_single_cancel_button(Callbacks.cancel_primary_action_callback)
         )
         return
-    success, response = await add_admin_to_db(username)
+    success, response = await add_or_update_user_to_admin(username)
+    telegram_tag = f"@{message.from_user.username}" if message.from_user.username else f"@{message.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
     if success:
-        await message.answer(text=response, reply_markup=main_actions_keyboard)
+        await message.answer(text=response, reply_markup=main_actions_keyboard(user.role))
         await state.clear()
     else:
         await message.answer(
@@ -103,28 +105,28 @@ async def process_admin_username(message: Message, state: FSMContext):
     await message.bot.delete_message(chat_id=message.chat.id, message_id=message.message_id - 1)
 
 
-@user.callback_query(F.data.startswith("admin_clicked"))
+@owner.callback_query(F.data.startswith("admin_clicked"))
 async def on_admin_clicked(callback: CallbackQuery):
     try:
-        admin_id = callback.data.split(":")[1]
+        user_id = callback.data.split(":")[1]
     except IndexError:
-        await callback.answer("Ошибка: админ не выбран!", show_alert=True)
+        await callback.answer("Ошибка: пользователь не выбран!", show_alert=True)
         return
-    admin = await get_admin_by_id(admin_id)
-    if not admin:
+    user = await get_user_by_id(user_id)
+    if not user or user.role != "admin":
         await callback.answer("Ошибка: админ не найден в базе данных!", show_alert=True)
-        print(f"Ошибка: админ не найден с id {admin_id}!")
+        print(f"Ошибка: админ не найден с id {user_id}!")
         return
     text = (
-        f"Username админа: {admin.username}\n"
+        f"Username админа: {user.telegram_tag}\n"
         "Админ имеет права на управление тегами (их добавление, изменение, архивацию и удаление)."
     )
     reply_markup = InlineKeyboardMarkup(
         inline_keyboard=[
             [
                 InlineKeyboardButton(
-                    text="🗑️ Удалить админа",
-                    callback_data=f"{Callbacks.admin_delete_callback}:{admin_id}",
+                    text="🗑️ Понизить до пользователя",
+                    callback_data=f"{Callbacks.admin_delete_callback}:{user_id}",
                 )
             ],
             [
@@ -139,20 +141,24 @@ async def on_admin_clicked(callback: CallbackQuery):
     await callback.answer("")
 
 
-@user.callback_query(F.data.startswith(Callbacks.admin_delete_callback))
+@owner.callback_query(F.data.startswith(Callbacks.admin_delete_callback))
 async def on_admin_delete_callback(callback: CallbackQuery):
     try:
-        admin_id = callback.data.split(":")[1]
+        user_id = callback.data.split(":")[1]
     except IndexError:
-        await callback.answer("Ошибка: админ не выбран!", show_alert=True)
+        await callback.answer("Ошибка: пользователь не выбран!", show_alert=True)
         return
-    success, response = await delete_admin_from_db(admin_id)
-    await callback.message.answer(text=response, reply_markup=main_actions_keyboard)
+    success, response = await demote_admin_to_user(user_id)
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
+    await callback.message.answer(text=response, reply_markup=main_actions_keyboard(user.role))
     await callback.message.delete()
     await callback.answer("")
 
 
-@user.callback_query(F.data == Callbacks.return_to_admin_list_callback)
+@owner.callback_query(F.data == Callbacks.return_to_admin_list_callback)
 async def on_return_to_admin_list(callback: CallbackQuery):
     await callback.message.edit_text(
         text="Выберите админа или создайте нового:",
@@ -161,3 +167,18 @@ async def on_return_to_admin_list(callback: CallbackQuery):
         ),
     )
     await callback.answer("")
+
+
+@owner.callback_query(F.data == Callbacks.cancel_primary_action_callback)
+async def on_cancel_primary_callback(callback: CallbackQuery, state: FSMContext):
+    telegram_tag = f"@{callback.from_user.username}" if callback.from_user.username else f"@{callback.from_user.id}"
+    user = await get_user_by_telegram_tag(telegram_tag)
+    if not user:
+        return
+    await callback.answer("")
+    await callback.message.delete()
+    await callback.message.answer(
+        text="Действие отменено. Выберите новое действие с помощью кнопок под клавиатурой.",
+        reply_markup=main_actions_keyboard(user.role),
+    )
+    await state.clear()
